training/requirements.txt
# Fine-tuning DINOv2 for card matching
# Install: pip install -r training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
chromadb>=1.0.0
rich>=13.0.0
Pillow>=10.0.0
numpy>=1.24.0
